+++ styles/globals.scss
@@ -15,12 +15,25 @@ a {
   box-sizing: border-box;
 }
 
-.deco-underline:hover {
+
+// TODO: 共通化
+.deco-underline-text:hover {
   text-decoration: underline; /* 下線 */
   text-decoration-thickness: 0.5em; /* 線の太さ */
   // text-decoration-color: rgba(255, 228, 0, 0.4); /* 線の色 */
   text-decoration-color: rgba(182,227,136,0.5);
   text-underline-offset: -0.2em; /* 線の位置。テキストに重なるようにやや上部にする */
   text-decoration-skip-ink: none; /* 下線と文字列が重なる部分でも下線が省略されない（線が途切れない） */
+}
+
+.deco-underline-element:hover {
+  .deco-underline-text {
+    text-decoration: underline; /* 下線 */
+    text-decoration-thickness: 0.5em; /* 線の太さ */
+    // text-decoration-color: rgba(255, 228, 0, 0.4); /* 線の色 */
+    text-decoration-color: rgba(182,227,136,0.5);
+    text-underline-offset: -0.2em; /* 線の位置。テキストに重なるようにやや上部にする */
+    text-decoration-skip-ink: none; /* 下線と文字列が重なる部分でも下線が省略されない（線が途切れない） */
+  }
 }
